rename tsx file to be capitalized
